# backend/openapi.yaml
openapi: 3.1.0
info:
  title: Graph Platform API
  version: 3.2.0
  description: API Graph Platform with Graph, FSM, Review, Ontology, RAG, Execution and Copilot engines.
servers:
  - url: http://localhost:3001/api
security:
  - bearerAuth: []
paths:
...
  /graphs:
    get: { operationId: listGraphs, security: [], responses: { '200': { $ref: '#/components/responses/Success' } } }
    post: { operationId: createGraph, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
  /graphs/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch: { operationId: updateGraph, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '200': { $ref: '#/components/responses/Success' } } }
    delete: { operationId: deleteGraph, responses: { '200': { $ref: '#/components/responses/Success' } } }
...
  /graphs/import:
    post: { operationId: importGraph, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
...
  /reviews:
    get: { operationId: listReviews, security: [], responses: { '200': { $ref: '#/components/responses/Success' } } }
    post: { operationId: createReview, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
  /reviews/{id}:
    parameters: [{ $ref: '#/components/parameters/Id' }]
    patch: { operationId: updateReview, requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '200': { $ref: '#/components/responses/Success' } } }
    delete: { operationId: deleteReview, responses: { '200': { $ref: '#/components/responses/Success' } } }
  /reviews/{id}/votes:
    post: { operationId: voteReview, parameters: [{ $ref: '#/components/parameters/Id' }], requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '201': { $ref: '#/components/responses/Success' } } }
  /reviews/{id}/transition:
    post: { operationId: transitionReview, parameters: [{ $ref: '#/components/parameters/Id' }], requestBody: { $ref: '#/components/requestBodies/JsonBody' }, responses: { '200': { $ref: '#/components/responses/Success' } } }
  /reviews/{id}/history:
    get: { operationId: reviewHistory, security: [], parameters: [{ $ref: '#/components/parameters/Id' }], responses: { '200': { $ref: '#/components/responses/Success' } } }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    Id:
      name: id
      in: path
      required: true
      schema: { type: string }
  requestBodies:
    JsonBody:
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  responses:
    Success:
      description: Successful response
      content:
        application/json:
          schema: { type: [object, array], additionalProperties: true }